notifications: optionally wait for membership screening
Optionally wait for Members to pass Membership Screening before sending
a join server channel message or join DM.

Currently, there is no good mechanism to wait for membership screening
to pass and then send a message (or template)--it has to be done with
some weird hacks, i.e. triggering on the system message (sent after
screening is past).

Signed-off-by: Luca Zeuch <l-zeuch@email.de>